Use newer base image
Use a version tagged base image (0.2.0-alpine) to build this image.
It uses a newer version of alpine, 3.13. This reduces the
vulnerabilities found by snyk.

Fix the starship install command as the newer starship release install
script does not work properly with zsh.

Remove redundant installation of curl that already ships with the base
image.